// src/programs/interface/programs.interface.ts
export interface ProgramsInterface {
  university: string;

  location: string;

  degreeType: string;

  subject: string;

  requiredDegree: string;

  tuitionFee: string;

  requiredGpa: string;

  languageOfStudy: string;

  englishLanguageTestScore: string;

  germanLanguageTestScore: string;

  greScore: string;

  appStartDateSummer: Date;

  appStartDateWinter: Date;

  appEndDateSummer: Date;

  appEndDateWinter: Date;

  programDuration: string;

  admissionType: string;

  courseWebsite: string;

  applyVia: string;

  ects: string;

  selectionProcedure: string;
}